requirements.txt
click==8.1.3
dnspython==2.2.1
Flask==2.2.2
flask-swagger-ui==4.11.1
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pymongo==4.3.2
python-decouple==3.6
Werkzeug==2.2.2
zipp==3.9.0
